+++ lib/editor/code_editor/auto_complete_list.dart
@@ -1,0 +1,145 @@
+import 'dart:math';
+
+import 'package:flutter/material.dart';
+import 'package:re_editor/re_editor.dart';
+import 'package:trayce/editor/code_editor/auto_scroll_list_view.dart';
+
+class DefaultCodeAutocompleteListView extends StatefulWidget implements PreferredSizeWidget {
+  static const double kItemHeight = 26;
+
+  final ValueNotifier<CodeAutocompleteEditingValue> notifier;
+  final ValueChanged<CodeAutocompleteResult> onSelected;
+
+  const DefaultCodeAutocompleteListView({super.key, required this.notifier, required this.onSelected});
+
+  @override
+  Size get preferredSize => Size(
+    250,
+    // 2 is border size
+    min(kItemHeight * notifier.value.prompts.length, 150) + 2,
+  );
+
+  @override
+  State<StatefulWidget> createState() => _DefaultCodeAutocompleteListViewState();
+}
+
+class _DefaultCodeAutocompleteListViewState extends State<DefaultCodeAutocompleteListView> {
+  @override
+  void initState() {
+    widget.notifier.addListener(_onValueChanged);
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    widget.notifier.removeListener(_onValueChanged);
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      constraints: BoxConstraints.loose(widget.preferredSize),
+      decoration: BoxDecoration(color: Colors.grey, borderRadius: BorderRadius.circular(6)),
+      child: AutoScrollListView(
+        controller: ScrollController(),
+        initialIndex: widget.notifier.value.index,
+        scrollDirection: Axis.vertical,
+        itemCount: widget.notifier.value.prompts.length,
+        itemBuilder: (context, index) {
+          final CodePrompt prompt = widget.notifier.value.prompts[index];
+          final BorderRadius radius = BorderRadius.only(
+            topLeft: index == 0 ? const Radius.circular(5) : Radius.zero,
+            topRight: index == 0 ? const Radius.circular(5) : Radius.zero,
+            bottomLeft: index == widget.notifier.value.prompts.length - 1 ? const Radius.circular(5) : Radius.zero,
+            bottomRight: index == widget.notifier.value.prompts.length - 1 ? const Radius.circular(5) : Radius.zero,
+          );
+          return InkWell(
+            borderRadius: radius,
+            onTap: () {
+              widget.onSelected(widget.notifier.value.copyWith(index: index).autocomplete);
+            },
+            child: Container(
+              width: double.infinity,
+              height: DefaultCodeAutocompleteListView.kItemHeight,
+              padding: const EdgeInsets.only(left: 5, right: 5),
+              alignment: Alignment.centerLeft,
+              decoration: BoxDecoration(
+                color: index == widget.notifier.value.index ? const Color.fromARGB(255, 255, 140, 0) : null,
+                borderRadius: radius,
+              ),
+              child: RichText(
+                text: prompt.createSpan(context, widget.notifier.value.input),
+                overflow: TextOverflow.ellipsis,
+                maxLines: 1,
+              ),
+            ),
+          );
+        },
+      ),
+    );
+  }
+
+  void _onValueChanged() {
+    setState(() {});
+  }
+}
+
+// =============================================================================
+// Extensions
+// =============================================================================
+
+extension _CodePromptExtension on CodePrompt {
+  InlineSpan createSpan(BuildContext context, String input) {
+    const TextStyle style = TextStyle();
+    final InlineSpan span = style.createSpan(
+      value: word,
+      anchor: input,
+      color: Colors.blue,
+      fontWeight: FontWeight.bold,
+    );
+    final CodePrompt prompt = this;
+    if (prompt is CodeFieldPrompt) {
+      return TextSpan(children: [span, TextSpan(text: ' ${prompt.type}', style: style.copyWith(color: Colors.cyan))]);
+    }
+    if (prompt is CodeFunctionPrompt) {
+      return TextSpan(
+        children: [span, TextSpan(text: '(...) -> ${prompt.type}', style: style.copyWith(color: Colors.cyan))],
+      );
+    }
+    return span;
+  }
+}
+
+extension _TextStyleExtension on TextStyle {
+  InlineSpan createSpan({
+    required String value,
+    required String anchor,
+    required Color color,
+    FontWeight? fontWeight,
+    bool casesensitive = false,
+  }) {
+    if (anchor.isEmpty) {
+      return TextSpan(text: value, style: this);
+    }
+    final int index;
+    if (casesensitive) {
+      index = value.indexOf(anchor);
+    } else {
+      index = value.toLowerCase().indexOf(anchor.toLowerCase());
+    }
+    if (index < 0) {
+      return TextSpan(text: value, style: this);
+    }
+    return TextSpan(
+      children: [
+        TextSpan(text: value.substring(0, index), style: this),
+        TextSpan(
+          text: value.substring(index, index + anchor.length),
+          style: copyWith(color: color, fontWeight: fontWeight),
+        ),
+        TextSpan(text: value.substring(index + anchor.length), style: this),
+      ],
+    );
+  }
+}
