+++ lib/editor/common/headers_table_read_only.dart
