+++ lib/editor/explorer/explorer.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../common/context_menu_style.dart';
+import '../../common/context_menu_style.dart';
 import 'explorer_style.dart';
 import 'file_node.dart';
