+++ lib/editor/explorer/file_node.dart
